[Refactor] Renamed IdentityInput to BaseInput
